Handle newer-ish versions of Proton-GE
Proton-GE tries to apply fixes automatically for certain games. When
called from the command line like this it throws a warning:

    WARN: Skipping fix execution. We are probably running an unit test.

This was causing CliWrap to throw an exception, but by capturing
StdOut and StdErr in StringBuilders ClipWrap seems much happier now,
so.... "Good enough"?

Also this commit improves the CliWrap declaration, improving readability
and reducing code duplication.